+++ playground/vite-plugin-cloudflare/workers/app.ts
@@ -14,7 +14,6 @@ declare module "react-router" {
 }
 
 const requestHandler = createRequestHandler(
-  // @ts-expect-error
   () => import("virtual:react-router/server-build"),
   import.meta.env.MODE
 );